requirements.txt
attrs==23.2.0
autolinkedin==1.6.3
browserjquery==0.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
cssselect==1.2.0
django-environ==0.11.2
exceptiongroup==1.2.1
fake-useragent==1.5.1
h11==0.14.0
idna==3.7
linkedin-scraper==2.11.2
lxml==5.2.1
outcome==1.3.0.post0
packaging==24.0
pyquery==2.0.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.19.0
selenium-stealth==1.0.6
seleniumtabs==0.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.0
wsproto==1.2.0
